Set debug configuration properties to default values.
miniupnp and libnatpmp MSVC project settings allow for incremental debug builds and edit and continue compilation. This causes a warning when compiling NetHelper in debug mode. Since incremental debug builds were disabled in NetHelper, MSVC was having to set both miniupnp and libnatpmp projects to disable incremental debug builds which was contrary to their settings.

https://docs.microsoft.com/en-us/cpp/build/reference/incremental-link-incrementally?view=vs-2017

Adding incremental debug builds means that some extra padding and 'jump thunks' (don't know what they are?) may be included in the compilation. We should watch and make sure this does not cause problems with how NetHelper hooks to Outpost 2. If problems are caused, we can disable it and go against the default setting. Incremental building is disabled in Release mode, so not a concern on final DLL.